calculate IBNR claims

given as input a data frame of insurance policies coming out from the function optimize_exposure(), and a table with IBNR factors, calculate the IBNR claims.

calculate_IBNR(.data, .ibnrtable, .ibnrfactor, .cy)

Arguments

.data
A data frame coming out from the function optimize_exposure(), or of the same type.

.ibnrtable
A table with two columns, one for the IBNR factors and one for the associated calendar year.

.ibnrfactor
the column of .ibnrtable with the IBNR factors.

.cy
the column of .ibnrtable with the corresponding calendar year of the IBNR factors.

Value

A new data frame with new columns:

IBNR_factor, a column that indicates the IBNR factor of the corresponding current year;

CLAIMS_IBNR, a column with the associated IBNR claim for each CY and policyholder (NOTE: if the current year in the original data frame is higher than the most recent current year in the IBNR table, this will produce an NA.)

Examples

data(lifepolicies)

lifepolicies %>%
  optimize_exposure(
  .polnr = Policy_id, .inceptiondate = Inception_date,
  .termdate = Term_date, .mainevent = Claim , .mainevent_date = Claim_date,
  .subevent = Lapse, .subevent_date = Lapse_date, .birthdate = Birth_date,
  .gender = Gender, .waitingperiod = Waiting_period
  ) %>%
  compute_expectedclaim(
    .table = SI2006, .age = AGE,
    .gender = GENDER, .qx = qx
  ) %>%
  calculate_IBNR(
    .ibnrtable = IBNR_factors_CQP,
    .ibnrfactor = IBNR_factor, .cy = CY
  )

#> Warning: There are policies which main event occurs in the waiting period: c(1046726, 1355512, 1807138, 2560830, 2759498, 3451651, 4017689, 4947699, 5141288, 5194342, 5461098, 6020022, 6518108, 7231957, 7263230, 7278494, 9169837, 9409256)

#> Warning: There are policies which sub event occurs in the waiting period: c(1425221, 2293556, 2713558, 3088353, 4809645, 5575964, 6504789, 6572333, 6793280, 7869755, 7950721, 9603585)

#> # A tibble: 28,514 x 27
#>    Policy_id Inception_date Term_date  Birth_date Gender    UY    CY
#>        <dbl> <date>         <date>     <date>     <chr>  <dbl> <dbl>
#>  1   1019640 2012-11-01     2016-11-01 1945-03-10 F       2012  2012
#>  2   1019640 2012-11-01     2016-11-01 1945-03-10 F       2012  2013
#>  3   1019640 2012-11-01     2016-11-01 1945-03-10 F       2012  2013
#>  4   1019640 2012-11-01     2016-11-01 1945-03-10 F       2012  2014
#>  5   1019640 2012-11-01     2016-11-01 1945-03-10 F       2012  2014
#>  6   1019640 2012-11-01     2016-11-01 1945-03-10 F       2012  2015
#>  7   1019640 2012-11-01     2016-11-01 1945-03-10 F       2012  2015
#>  8   1019640 2012-11-01     2016-11-01 1945-03-10 F       2012  2016
#>  9   1019640 2012-11-01     2016-11-01 1945-03-10 F       2012  2016
#> 10   1019640 2012-11-01     2016-11-01 1945-03-10 F       2012  2017
#> # ... with 28,504 more rows, and 20 more variables: policy_duration <dbl>,
#> #   Claim_date <date>, Claim <dbl>, Lapse_date <date>, Lapse <dbl>,
#> #   inception_valuation_date <date>, end_valuation_date <date>, AGE0 <dbl>,
#> #   AGE1 <dbl>, DAY_AGE0 <dbl>, DAY_AGE1 <dbl>, EXPOSURE_YEAR_AGE0 <dbl>,
#> #   EXPOSURE_YEAR_AGE1 <dbl>, EXPOSURE_IN_YEAR <dbl>, qx.x <dbl>, qx.y <dbl>,
#> #   EXPECTED_CLAIM_AGE0 <dbl>, EXPECTED_CLAIM_AGE1 <dbl>, IBNR_factor <dbl>,
#> #   CLAIMS_IBNR <dbl>