Compute the expected claims

Given as input a data frame of insurance policies coming out from optimize_exposure(), it returns a new data frame with the associated expected claim for each age of the policyhoder.

compute_expectedclaim(.data, .table, .age, .gender, .qx)

Arguments

.data
data frame in the format provided by optimize_exposure().

.table
a table with the probabilities qx.

.age
the .table's column with the age associated to qx.

.gender
the .table's column with the gender associated with each qx.

.qx
the .table's column with the probabilities qx (note that it has to be per 1000).

Value

A new data frame with 4 new columns:

qx.x indicates the probability of dying between AGE0 and AGE1 for the policyholder;

qx.y indicates the probability of dying between AGE1 and AGE1+ 1 for the policyolder;

EXPECTED_CLAIM_AGE0 is the expected claim for the policyholder at AGE0, computed as the exposure in year when he/she has AGE0 * (qx.x/ 1000);

EXPECTED_CLAIM_AGE1 is the expected claim for the policyholder at AGE1, computed as the exposure in year when he/she has AGE1 * (qx.y/ 1000).

Examples

data(lifepolicies)

lifepolicies %>%
  optimize_exposure(
  .polnr = Policy_id, .inceptiondate = Inception_date,
  .termdate = Term_date, .mainevent = Claim , .mainevent_date = Claim_date,
  .subevent = Lapse, .subevent_date = Lapse_date, .birthdate = Birth_date,
  .gender = Gender, .waitingperiod = Waiting_period
  ) %>%
  compute_expectedclaim(
    .table = SI2006, .age = AGE,
    .gender = GENDER, .qx = qx
  )

#> # A tibble: 28,514 x 25
#>    Policy_id Inception_date Term_date  Birth_date Gender    UY    CY
#>        <dbl> <date>         <date>     <date>     <chr>  <dbl> <dbl>
#>  1   1019640 2012-11-01     2016-11-01 1945-03-10 F       2012  2012
#>  2   1019640 2012-11-01     2016-11-01 1945-03-10 F       2012  2013
#>  3   1019640 2012-11-01     2016-11-01 1945-03-10 F       2012  2013
#>  4   1019640 2012-11-01     2016-11-01 1945-03-10 F       2012  2014
#>  5   1019640 2012-11-01     2016-11-01 1945-03-10 F       2012  2014
#>  6   1019640 2012-11-01     2016-11-01 1945-03-10 F       2012  2015
#>  7   1019640 2012-11-01     2016-11-01 1945-03-10 F       2012  2015
#>  8   1019640 2012-11-01     2016-11-01 1945-03-10 F       2012  2016
#>  9   1019640 2012-11-01     2016-11-01 1945-03-10 F       2012  2016
#> 10   1019640 2012-11-01     2016-11-01 1945-03-10 F       2012  2017
#> # ... with 28,504 more rows, and 18 more variables: policy_duration <dbl>,
#> #   Claim_date <date>, Claim <dbl>, Lapse_date <date>, Lapse <dbl>,
#> #   inception_valuation_date <date>, end_valuation_date <date>, AGE0 <dbl>,
#> #   AGE1 <dbl>, DAY_AGE0 <dbl>, DAY_AGE1 <dbl>, EXPOSURE_YEAR_AGE0 <dbl>,
#> #   EXPOSURE_YEAR_AGE1 <dbl>, EXPOSURE_IN_YEAR <dbl>, qx.x <dbl>, qx.y <dbl>,
#> #   EXPECTED_CLAIM_AGE0 <dbl>, EXPECTED_CLAIM_AGE1 <dbl>