Life policies portfolio (n x 2 rows for each policy).

Simulated life insurance policies portfolio made up of 2,500 policies. Note that this type of dataset is suited to be passed as input in the function optimize_exposure(). This because it has n x 2 rows for each policy, where n is the number of years the policy is in force.

lifepolicies

Format

A data frame with variables:

Policy_id

policy number identifier, unique for each policy.

Inception_date

signing date of the contract.

Term_date

ending date of the contract.

Birth_date

date of birth of the policyolder.

Gender

policyholder's gender.

UY

underwriting year.

CY

current year of evaluation.

Claim_date

date when claims occurred, if any, otherwise 2099-12-31.

Claim

1 if the year of Claim_date is equal to CY, 0 otherwise.

Lapse_date

date when policy lapsed, if so, otherwise 2099-12-31.

Lapse

1 if the year of Lapse_date is equal to CY, 0 otherwise.

Examples

data(lifepolicies)
head(lifepolicies)

#> # A tibble: 6 x 11
#>   Policy_id Inception_date Term_date  Birth_date Gender    UY    CY Claim_date
#>       <dbl> <date>         <date>     <date>     <chr>  <dbl> <int> <date>    
#> 1   1000883 2012-04-01     2020-04-01 1949-02-10 F       2012  2012 2099-12-31
#> 2   1000883 2012-04-01     2020-04-01 1949-02-10 F       2012  2013 2099-12-31
#> 3   1000883 2012-04-01     2020-04-01 1949-02-10 F       2012  2014 2099-12-31
#> 4   1000883 2012-04-01     2020-04-01 1949-02-10 F       2012  2015 2099-12-31
#> 5   1000883 2012-04-01     2020-04-01 1949-02-10 F       2012  2016 2099-12-31
#> 6   1000883 2012-04-01     2020-04-01 1949-02-10 F       2012  2017 2099-12-31
#> # ... with 3 more variables: Claim <dbl>, Lapse_date <date>, Lapse <dbl>