Compute optimal exposure

The idea behind this function is to compute in an optimal way the exposure period for each policyhoder and for each year in which the contract is in force, by distinguishing each calendar year in before and after his/her birthday. Doing so, for a given policyholder and for each year we get the exposure he/she is exposed with age x and the exposure he/she is exposed at age x+1.

optimize_exposure(
  .data,
  .polnr,
  .inceptiondate,
  .termdate,
  .mainevent,
  .mainevent_date,
  .subevent,
  .subevent_date,
  .birthdate,
  .gender,
  .waitingperiod = 0
)

Arguments

.data
A data frame of insurance policies that has a number of rows for each policy equal to: year in force of the policy x 2. This type of format is required to compute in an optimal way the exposure in each calendar year for each age of the policyholder.

.polnr
A column of .data with the policy identification numbers.

.inceptiondate
A column of .data with the starting date of the policies.

.termdate
A column of .data with the expiry date of the policies.

.mainevent
A column of .data with 1 if the main event (the one for which perform the monitoring) occurred, 0 otherwise.

.mainevent_date
A column of .data with the date of the event (claim or lapse) for which the monitoring has to be performed. For policy with no claim/lapse the date should be 2099-12-31.

.subevent
A column of .data with 1 if the sub event (the one for which do not perform the monitoring) occurred, 0 otherwise.

.subevent_date
A column of .data with the date of the subordinate event (claim or lapse). In this case, for this event the monitoring is not performed. For policies with no claim/lapse the date should be 2099-12-31.

.birthdate
A column of .data with the birthday of the policyholders.

.gender
A column of .data with the gender of the policyholders.

.waitingperiod
A column with the waiting period (in months) for the policy.

Value

The modified data frame with some new columns, in particular:

inception_valuation_date indicates the date from which to start to compute the exposure, it can be the inception date, the renewal date or the first of January of the current year;

end_valuation_date indicates the ending date to take into account in the computation of the exposure, it can be the last day of the current year, the claim/lapse date or the renewal date (this because in the following row the inception_valuation_date will be the same);

AGE0 is the age of the policyholder at the inception valuation date;

AGE1 is the possible age of the policyholder at the end valuation date;

DAY_AGE0 indicates the day of exposure of the policyholder with AGE0(before his/her anniversary) in the valuation period;

DAY_AGE1 indicates the day of exposure of the policyholder with AGE1 (after his/her anniversary) in the valuation period;

EXPOSURE_YEAR_AGE0 is the exposure of the policyholder in the current year when he/she has AGE0, (it is computed as number days exposed with age0 / total number of days in the current year);

EXPOSURE_YEAR_AGE1 is the exposure of the policyholder in the current year when he/she has AGE1, (it is computed as number days exposed with age1 / total number of days in the current year);

EXPOSURE_IN_YEAR is the total exposure in the valuation period, the sum of the previous two exposures, it is 1 by definition when a main event (claim/lapse) occurs.

Examples

data(lifepolicies)
lifepolicies %>%
  optimize_exposure(
  .polnr = Policy_id, .inceptiondate = Inception_date,
  .termdate = Term_date, .mainevent = Claim , .mainevent_date = Claim_date,
  .subevent = Lapse, .subevent_date = Lapse_date, .birthdate = Birth_date,
  .gender = Gender, .waitingperiod = Waiting_period
  )

#> Warning: There are policies which main event occurs in the waiting period: c(1046726, 1355512, 1807138, 2560830, 2759498, 3451651, 4017689, 4947699, 5141288, 5194342, 5461098, 6020022, 6518108, 7231957, 7263230, 7278494, 9169837, 9409256)

#> Warning: There are policies which sub event occurs in the waiting period: c(1425221, 2293556, 2713558, 3088353, 4809645, 5575964, 6504789, 6572333, 6793280, 7869755, 7950721, 9603585)

#> # A tibble: 28,514 x 21
#>    Policy_id Inception_date Term_date  Birth_date Gender    UY    CY
#>        <dbl> <date>         <date>     <date>     <chr>  <dbl> <dbl>
#>  1   1019640 2012-11-01     2016-11-01 1945-03-10 F       2012  2012
#>  2   1019640 2012-11-01     2016-11-01 1945-03-10 F       2012  2013
#>  3   1019640 2012-11-01     2016-11-01 1945-03-10 F       2012  2013
#>  4   1019640 2012-11-01     2016-11-01 1945-03-10 F       2012  2014
#>  5   1019640 2012-11-01     2016-11-01 1945-03-10 F       2012  2014
#>  6   1019640 2012-11-01     2016-11-01 1945-03-10 F       2012  2015
#>  7   1019640 2012-11-01     2016-11-01 1945-03-10 F       2012  2015
#>  8   1019640 2012-11-01     2016-11-01 1945-03-10 F       2012  2016
#>  9   1019640 2012-11-01     2016-11-01 1945-03-10 F       2012  2016
#> 10   1019640 2012-11-01     2016-11-01 1945-03-10 F       2012  2017
#> # ... with 28,504 more rows, and 14 more variables: policy_duration <dbl>,
#> #   Claim_date <date>, Claim <dbl>, Lapse_date <date>, Lapse <dbl>,
#> #   inception_valuation_date <date>, end_valuation_date <date>, AGE0 <dbl>,
#> #   AGE1 <dbl>, DAY_AGE0 <dbl>, DAY_AGE1 <dbl>, EXPOSURE_YEAR_AGE0 <dbl>,
#> #   EXPOSURE_YEAR_AGE1 <dbl>, EXPOSURE_IN_YEAR <dbl>