Compute optimal exposure

The idea behind this function is to compute in an optimal way the exposure period for each policyhoder and for each year in which the contract is in force, by distinguishing each calendar year in before and after his/her birthday. Doing so, for a given policyholder and for each year we get the exposure he/she is exposed with age x and the exposire he/she is exposed at age x+1.

optimize_exposure(
  .data,
  .polnr,
  .inceptiondate,
  .termdate,
  .claimdate,
  .lapsedate,
  .birthdate,
  .gender
)

Arguments

.data
A data frame of insurance policies that has a number of rows for each policy equal to: year in force of the policy x 2. This type of format is required to compute in an optimal way the exposure in each calendar year for each age of the policyholder.

.polnr
A column of .data with the policy identification numbers.

.inceptiondate
A column of .data with the starting date of the policies.

.termdate
A column of .data with the expiry date of the policies.

.claimdate
A column of .data with the date of the claim, if any, otherwise 2099-12-31.

.lapsedate
A column of .data with the date of the lapse, if any, otherwise 2099-12-31.

.birthdate
A column of .data with the birthday of the policyholders.

.gender
A column of .data with the gender of the policyholders.

Value

The modified data frame with some new columns, in particular:

inception_valuation_date indicates the date from which to start to compute the exposure, it can be the inception date, the renewal date or the first of January of the current year;

end_valuation_date indicates the ending date to take into account in the computation of the exposure, it can be the last day of the current year, the claim/lapse date or the renewal date (this because in the following row the inception_valuation_date will be the same);

AGE0 is the age of the policyholder at the inception valuation date;

AGE1 is the possible age of the policyholder at the end valuation date;

DAY_AGE0 indicates the day of exposure of the policyholder with AGE0(before his/her anniversary) in the valuation period;

DAY_AGE1 indicates the day of exposure of the policyholder with AGE1 (after his/her anniversary) in the valuation period;

EXPOSURE_YEAR_AGE0 is the exposure of the policyholder in the current year when he/she has AGE0, (it is computed as number days exposed with age0 / total number of days in the current year);

EXPOSURE_YEAR_AGE1 is the exposure of the policyholder in the current year when he/she has AGE1, (it is computed as number days exposed with age1 / total number of days in the current year);

EXPOSURE_IN_YEAR is the total exposure in the valuation period, the sum of the previous two exposures, it is 1 by definition when a claim occurs;

EXPOSURE_IN_YEAR_lapse is the total exposure in the valuation period, the sum of the previous two exposures, it is 1 by definition when a policy lapses.

Examples

data(lifepolicies)
lifepolicies %>%
  optimize_exposure(
    .polnr = Policy_id, .inceptiondate = Inception_date,
    .termdate = Term_date, .claimdate = Claim_date,
    .lapsedate = Lapse_date, .birthdate = Birth_date,
    .gender = Gender
  )

#> # A tibble: 29,743 x 22
#>    Policy_id Inception_date Term_date  Birth_date Gender    UY    CY
#>        <dbl> <date>         <date>     <date>     <chr>  <dbl> <dbl>
#>  1   1000883 2012-04-01     2020-04-01 1949-02-10 F       2012  2012
#>  2   1000883 2012-04-01     2020-04-01 1949-02-10 F       2012  2013
#>  3   1000883 2012-04-01     2020-04-01 1949-02-10 F       2012  2013
#>  4   1000883 2012-04-01     2020-04-01 1949-02-10 F       2012  2014
#>  5   1000883 2012-04-01     2020-04-01 1949-02-10 F       2012  2014
#>  6   1000883 2012-04-01     2020-04-01 1949-02-10 F       2012  2015
#>  7   1000883 2012-04-01     2020-04-01 1949-02-10 F       2012  2015
#>  8   1000883 2012-04-01     2020-04-01 1949-02-10 F       2012  2016
#>  9   1000883 2012-04-01     2020-04-01 1949-02-10 F       2012  2016
#> 10   1000883 2012-04-01     2020-04-01 1949-02-10 F       2012  2017
#> # ... with 29,733 more rows, and 15 more variables: policy_duration <dbl>,
#> #   Claim_date <date>, Claim <dbl>, Lapse_date <date>, Lapse <dbl>,
#> #   inception_valuation_date <date>, end_valuation_date <date>, AGE0 <dbl>,
#> #   AGE1 <dbl>, DAY_AGE0 <dbl>, DAY_AGE1 <dbl>, EXPOSURE_YEAR_AGE0 <dbl>,
#> #   EXPOSURE_YEAR_AGE1 <dbl>, EXPOSURE_IN_YEAR <dbl>,
#> #   EXPOSURE_IN_YEAR_LAPSE <dbl>